DDA-69: Extract tools from grid lens
