+++ src/react/features/lenses/grid/GridLens.js
@@ -1,12 +1,12 @@
 import React from 'react'
 
 import { useGridLens } from './useGridLens'
-import { useTools } from './tools/useTools'
+import { useTools } from '../../tools/useTools'
 
 import Grid from '../../../components/Grid'
-import CreateTool from './tools/create/CreateTool'
-import DeleteTool from './tools/delete/DeleteTool'
-import UpdateTool from './tools/update/UpdateTool'
+import CreateTool from '../../tools/create/CreateTool'
+import DeleteTool from '../../tools/delete/DeleteTool'
+import UpdateTool from '../../tools/update/UpdateTool'
 
 const GridLens = () => {
   const { diory, diorys } = useGridLens()
