+++ src/react/features/tools/delete/DeleteTool.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { useButtons } from '../../../../buttons'
+import { useButtons } from '../../buttons'
 
 import { buttons } from './buttons'
