+++ src/react/features/tools/delete/buttons.js
