+++ src/react/features/tools/update/UpdateView.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { action } from '@storybook/addon-actions'
-import { StoreProvider } from '../../../../../store/StoreContext'
+import { StoreProvider } from '../../../store/StoreContext'
 
 import UpdateView from './UpdateView'
