+++ src/react/features/tools/update/updateTool.feature
